move app to components
